+++ packages/react-components/react-headless-components-preview/library/src/components/Drawer/DrawerHeaderNavigation/useDrawerHeaderNavigation.ts
@@ -1,0 +1,8 @@
+'use client';
+
+import { useDrawerHeaderNavigation_unstable } from '@fluentui/react-drawer';
+
+/**
+ * Returns the state for a DrawerHeaderNavigation component, given its props and ref.
+ */
+export const useDrawerHeaderNavigation = useDrawerHeaderNavigation_unstable;