+++ packages/react-components/react-headless-components-preview/stories/rit.config.js
@@ -1,0 +1,15 @@
+// @ts-check
+
+/** @type {import('@fluentui/react-integration-tester').Config} */
+const config = {
+  react: {
+    17: {
+      runConfig: { 'type-check': { configPath: 'tsconfig.lib.json' } },
+    },
+    18: {
+      runConfig: { 'type-check': { configPath: 'tsconfig.lib.json' } },
+    },
+  },
+};
+
+module.exports = config;