+++ packages/react-components/react-headless-components-preview/stories/src/Breadcrumb/index.stories.tsx
@@ -3,7 +3,7 @@ import {
   BreadcrumbItem,
   BreadcrumbButton,
   BreadcrumbDivider,
-} from '@fluentui/react-headless-components-preview';
+} from '@fluentui/react-headless-components-preview/breadcrumb';
 
 import descriptionMd from './BreadcrumbDescription.md';