+++ packages/react-components/react-headless-components-preview/stories/src/Link/index.stories.tsx
@@ -1,4 +1,4 @@
-import { Link } from '@fluentui/react-headless-components-preview';
+import { Link } from '@fluentui/react-headless-components-preview/link';
 
 import descriptionMd from './LinkDescription.md';