+++ packages/react-components/react-headless-components-preview/stories/src/TabList/index.stories.tsx
@@ -1,4 +1,4 @@
-import { TabList } from '@fluentui/react-headless-components-preview';
+import { TabList } from '@fluentui/react-headless-components-preview/tab-list';
 
 import descriptionMd from './TabListDescription.md';