+++ packages/react-components/react-headless-components-preview/stories/src/Toolbar/ToolbarDefault.stories.tsx
@@ -6,7 +6,7 @@ import {
   ToolbarGroup,
   ToolbarRadioGroup,
   ToolbarToggleButton,
-} from '@fluentui/react-headless-components-preview';
+} from '@fluentui/react-headless-components-preview/toolbar';
 import {
   CutRegular,
   CopyRegular,