+++ packages/react-components/react-link/library/CHANGELOG.md
@@ -1,8 +1,19 @@
 # Change Log - @fluentui/react-link
 
-This log was last generated on Thu, 26 Mar 2026 08:12:50 GMT and should not be manually modified.
+This log was last generated on Thu, 23 Apr 2026 14:21:12 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## [9.8.1](https://github.com/microsoft/fluentui/tree/@fluentui/react-link_v9.8.1)
+
+Thu, 23 Apr 2026 14:21:12 GMT 
+[Compare changes](https://github.com/microsoft/fluentui/compare/@fluentui/react-link_v9.8.0..@fluentui/react-link_v9.8.1)
+
+### Patches
+
+- Bump @fluentui/react-jsx-runtime to v9.4.2 ([PR #36035](https://github.com/microsoft/fluentui/pull/36035) by beachball)
+- Bump @fluentui/react-tabster to v9.26.14 ([PR #36035](https://github.com/microsoft/fluentui/pull/36035) by beachball)
+- Bump @fluentui/react-utilities to v9.26.3 ([PR #36035](https://github.com/microsoft/fluentui/pull/36035) by beachball)
 
 ## [9.8.0](https://github.com/microsoft/fluentui/tree/@fluentui/react-link_v9.8.0)